test: move tests which use mockserver to src/test
src/it directory is reserved for tests that interact with the actual
Weaviate instance. They're usually slower and more expensive to set up.
